backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlmodel>=0.0.14
pydantic>=2.4.0
python-multipart>=0.0.6
pyyaml>=6.0
